fix: error to capture consoles and exposed types
