+++ .npmignore
@@ -5,3 +5,8 @@ karma.conf.js
 rollup.config.js
 tsconfig.json
 test.tsconfig.json
+jest.config.js
+tsconfig-build.json
+.prettierrc.js
+.eslintrc.js
+.eslintignore
