+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ra-mocha-as-promised",
-  "version": "1.1.1",
+  "version": "1.1.7",
   "description": "A mocha.js and chai wrapper that delivers a promise-based API to get test results instead of only printing.",
   "main": "./dist/index.js",
   "types": "dist/index.d.ts",
@@ -8,7 +8,9 @@
     "start": "node ./dist/index.js",
     "dev": "NODE_ENV=development ts-node -r tsconfig-paths/register ./src/index.ts",
     "build": "rm -rf ./dist && ttsc",
-    "test": "jest --silent --passWithNoTests --noStackTrace --runInBand"
+    "test": "jest --silent --passWithNoTests --noStackTrace --runInBand",
+    "publish": "npm run build && npm publish",
+    "pack": "npm run build && npm pack"
   },
   "repository": {
     "type": "git",
